Add files via upload
This is a crime records management system with 2 logins. The victim and police login. Victim can login through given ID and password to check the status of their case. Police can login and view all the information of the crimes that took place around them. They are allowed to add, delete or edit the information. The system contains information about the crime spot , date and time, the policed assigned to that crime, status of crime. It also contains the suspects and the reason to suspect which might help to analyze the case. If caught, the details of the criminal, his crime history and court too is present.